QuickFilter component to prevent React reload of everything on the page every keydown
