+++ src/components/movies/MovieList.jsx
@@ -1,8 +1,8 @@
 import { useCallback, useEffect, useState } from 'react'
 import MovieShowtimeSummary from './MovieShowtimeSummary'
+import QuickFilter from './QuickFilter'
 
 const MovieList = ({ movies, isLoading, className }) => {
-  const [query, setQuery] = useState('')
   const [currentQuery, setCurrentQuery] = useState('')
   const [filteredMovies, setFilteredMovies] = useState([])
 
@@ -21,34 +21,13 @@ const MovieList = ({ movies, isLoading, className }) => {
     updateFilteredMovies(currentQuery)
   }, [updateFilteredMovies, currentQuery])
 
-  const handleSubmit = (e) => {
-    e.preventDefault()
-    setCurrentQuery(query)
-  }
-
-  const handleQueryChange = (e) => {
-    setQuery(e.target.value)
-  }
-
   if (isLoading) {
     return <div className='centre loading'></div>
   }
 
   return (
     <div className={['movies', className].join(' ')}>
-      <form onSubmit={handleSubmit}>
-        <h3>Quick filter</h3>
-        <label htmlFor='name'>By name</label>{' '}
-        <input
-          type='text'
-          id='name'
-          name='name'
-          value={query}
-          placeholder="e.g. 'Batman'"
-          onChange={handleQueryChange}
-        />
-        <input type='submit' value='Run' />
-      </form>
+      <QuickFilter setCurrentQuery={setCurrentQuery} />
       {filteredMovies.map((props) => (
         <MovieShowtimeSummary key={props.movie._id} {...props} />
       ))}
